feat: Check folder permissions feat: Add internal website for configuring sites feat: Add Not Found request handler feat: Support starting all sites for console app feat: Add scripts

+++ CFWebServerCommon/WebRequestHandlers/StaticResourceDeleteWebRequestHandler.cs
@@ -1,4 +1,6 @@
-﻿using CFWebServer.Interfaces;
+﻿using CFWebServer.Constants;
+using CFWebServer.Enums;
+using CFWebServer.Interfaces;
 using CFWebServer.Models;
 using CFWebServer.Utilities;
 using System.Net;
@@ -16,39 +18,49 @@ namespace CFWebServer.WebRequestHandlers
             
         }
 
-        public bool CanHandle(RequestContext requestContext)
-        {
-            return requestContext.Request.HttpMethod == "DELETE";
-        }
+        public string Name => WebRequestHandlerNames.StaticResourceDelete;
+
+        //public bool CanHandle(RequestContext requestContext)
+        //{
+        //    return requestContext.Request.HttpMethod == "DELETE";
+        //}
 
         public async Task HandleAsync(RequestContext requestContext)
         {
-            if (!CanHandle(requestContext))
-            {
-                throw new ArgumentException("Unable to handle request");
-            }
+            //if (!CanHandle(requestContext))
+            //{
+            //    throw new ArgumentException("Unable to handle request");
+            //}
 
             var relativePath = requestContext.Request.Url.AbsolutePath;
 
-            // Get local path
-            var localResourcePath = GetResourceLocalPath(relativePath);
-            
             var response = requestContext.Response;
-            if (File.Exists(localResourcePath))
+
+            if (IsActionAllowedForFolderPermission(HttpUtilities.GetUrlWithoutLastElement(relativePath), FolderPermissions.Write))
             {
-                File.Delete(localResourcePath);
+                // Get local path
+                var localResourcePath = GetResourceLocalPath(relativePath);                
+                if (File.Exists(localResourcePath))
+                {
+                    File.Delete(localResourcePath);
+                }
+                else
+                {
+                    response.StatusCode = (int)HttpStatusCode.NotFound;
+                }
+                response.Close();
+
+                // Remove cache file if exists
+                var cacheFile = _fileCacheService.Get(relativePath);
+                if (cacheFile != null)
+                {
+                    _fileCacheService.Remove(relativePath);
+                }
             }
             else
             {
-                response.StatusCode = (int)HttpStatusCode.NotFound;
-            }
-            response.Close();
-
-            // Remove cache file if exists
-            var cacheFile = _fileCacheService.Get(relativePath);
-            if (cacheFile != null)
-            {
-                _fileCacheService.Remove(relativePath);
+                response.StatusCode = (int)HttpStatusCode.Unauthorized;
+                response.Close();
             }
         }
     }
